backend/process_pdf_function/requirements.txt
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.2
azure-storage-blob==12.14.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
dnspython==2.2.1
et-xmlfile==1.1.0
idna==3.4
isodate==0.6.1
loguru==0.6.0
msrest==0.7.1
numpy==1.24.1
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.230105
pinecone-client==2.1.0
pycparser==2.21
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tqdm==4.64.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
azure-cosmos==4.3.0
uuid==1.30
azure-functions-durable
stripe